Rename feature _components directories
