+++ app/features/bookmarks/__tests__/Bookmarks.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import { SettingsProvider } from '@/app/providers/SettingsContext';
-import BookmarkedVersesList from '@/app/features/bookmarks/_components/BookmarkedVersesList';
+import BookmarkedVersesList from '@/app/features/bookmarks/components/BookmarkedVersesList';
 import BookmarksPage from '@/app/features/bookmarks/page';
 import * as api from '@/lib/api';
 import { Verse } from '@/types';
