+++ app/features/surah/[surahId]/components/CollapsibleSection.tsx
@@ -1,4 +1,4 @@
-// app/surah/[surahId]/_components/CollapsibleSection.tsx
+// app/features/surah/[surahId]/components/CollapsibleSection.tsx
 'use client';
 import { FaChevronDown } from '@/app/components/shared/SvgIcons';
