+++ app/features/surah/[surahId]/components/TafsirPanel.tsx
