+++ app/features/surah/[surahId]/components/TranslationPanel.tsx
