+++ app/features/surah/[surahId]/components/WordLanguagePanel.tsx
