+++ app/features/surah/__tests__/SettingsSidebar.test.tsx
@@ -2,9 +2,9 @@ import React, { useState } from 'react';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import Header from '@/app/components/shared/Header';
-import { SettingsSidebar } from '@/app/features/surah/[surahId]/_components/SettingsSidebar';
-import { WordLanguagePanel } from '@/app/features/surah/[surahId]/_components/WordLanguagePanel';
-import { TranslationPanel } from '@/app/features/surah/[surahId]/_components/TranslationPanel';
+import { SettingsSidebar } from '@/app/features/surah/[surahId]/components/SettingsSidebar';
+import { WordLanguagePanel } from '@/app/features/surah/[surahId]/components/WordLanguagePanel';
+import { TranslationPanel } from '@/app/features/surah/[surahId]/components/TranslationPanel';
 import { SettingsProvider } from '@/app/providers/SettingsContext';
 import { SidebarProvider } from '@/app/providers/SidebarContext';
 import { ThemeProvider } from '@/app/providers/ThemeContext';
