+++ app/features/tafsir/[surahId]/[ayahId]/components/TafsirTabs.tsx
