+++ app/features/tafsir/[surahId]/[ayahId]/components/TafsirVerse.tsx
