+++ app/features/tafsir/[surahId]/[ayahId]/components/VerseCard.tsx
