+++ app/features/tafsir/[surahId]/[ayahId]/page.tsx
@@ -2,12 +2,12 @@
 import React, { useState, useMemo, useEffect } from 'react';
 import { useParams, useRouter } from 'next/navigation';
 import { useTranslation } from 'react-i18next';
-import { Verse as VerseComponent } from '@/app/features/surah/[surahId]/_components/Verse';
-import { SettingsSidebar } from '@/app/features/surah/[surahId]/_components/SettingsSidebar';
-import { TranslationPanel } from '@/app/features/surah/[surahId]/_components/TranslationPanel';
-import { TafsirPanel } from '@/app/features/surah/[surahId]/_components/TafsirPanel';
-import { WordLanguagePanel } from '@/app/features/surah/[surahId]/_components/WordLanguagePanel';
-import TafsirTabs from './_components/TafsirTabs';
+import { Verse as VerseComponent } from '@/app/features/surah/[surahId]/components/Verse';
+import { SettingsSidebar } from '@/app/features/surah/[surahId]/components/SettingsSidebar';
+import { TranslationPanel } from '@/app/features/surah/[surahId]/components/TranslationPanel';
+import { TafsirPanel } from '@/app/features/surah/[surahId]/components/TafsirPanel';
+import { WordLanguagePanel } from '@/app/features/surah/[surahId]/components/WordLanguagePanel';
+import TafsirTabs from './components/TafsirTabs';
 import {
   getVersesByChapter,
   getTranslations,
