+++ app/features/tafsir/__tests__/TafsirVerseCard.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import VerseCard from '@/app/features/tafsir/[surahId]/[ayahId]/_components/VerseCard';
+import VerseCard from '@/app/features/tafsir/[surahId]/[ayahId]/components/VerseCard';
 import { SettingsProvider } from '@/app/providers/SettingsContext';
 import { AudioProvider } from '@/app/features/player/context/AudioContext';
 import { Verse } from '@/types';
